(MAINT) Confine aws-sdk gem to bundler group
It's currently an optinal engine, and it should specify an optional
dependency. This also makes the spec test for the EC2 engine option
depending on whether or not the aws-sdk is available.

When this engine is declared a prime-time engine, the dependency on
AWS-SDK should be moved directly into the Gemspec and this confinement
should be removed.